+++ NEWS.md
@@ -1,3 +1,17 @@
+# exdqlm 0.1.2
+- Fixes and general improvements
+  - fixed a bug in the forecast plotting routine
+  - dqlm.ind = TRUE automatically sets gam.init = 0 & fix.gamma = TRUE (implemented in check_logics)
+  - added a check that gam.init is in the appropriate range in exdqlmISVB and exdqlmMCMC
+- Input changes
+  - added parameter to specify the percentage of the CrIs in compPlot, exdqlmPlot and exdqlmForecast
+- Return changes
+  - kt added to transfn_exdqlmISVB return
+- Dataset changes
+  - Niño 3.4 dataset added
+  - BTflow dataset updated
+  - monELI dataset removed
+
 # exdqlm 0.1.1
 - Documentation updates
   - added author (year) to citation in descriptions